[ADD] sale_purchase_stock: purchase count from sale order
This commit simply moves some code from stock_dropshipping to a
new bridge module between sale, stock and purchase.
The goal is to have a correct count of purchase created from a
sales order, either with a stockable product or a service.

To do so, we need that code to link purchase line created in a
MTO/Buy selling scenario, with its origin sale line.

We have to create this new bridge module above those 3, since they
don't share a common dependency in their flow (RIP procurement).

Task-1906481
