+++ addons/sale_purchase/views/sale_views.xml
@@ -17,7 +17,7 @@
                </button>
             </xpath>
         </field>
-        <field name="active" eval="False"/>
+        <field name="active" eval="True"/>
     </record>
 
 </odoo>
