+++ addons/stock_dropshipping/models/purchase.py
@@ -13,13 +13,3 @@ class PurchaseOrderLine(models.Model):
         for re in res:
             re['sale_line_id'] = self.sale_line_id.id
         return res
-
-
-class StockRule(models.Model):
-    _inherit = 'stock.rule'
-
-    @api.model
-    def _prepare_purchase_order_line(self, product_id, product_qty, product_uom, values, po, supplier):
-        res = super(StockRule, self)._prepare_purchase_order_line(product_id, product_qty, product_uom, values, po, supplier)
-        res['sale_line_id'] = values.get('sale_line_id', False)
-        return res
